Corrected pose output and entity attributedisplayattribute input/output

+++ api/src/main/java/net/donnypz/displayentityutils/utils/DisplayEntities/PacketDisplayEntityPart.java
@@ -555,7 +555,7 @@ public class PacketDisplayEntityPart extends ActivePart implements Packeted{
     @Override
     public void setMannequinScale(double scale) {
         if (type != SpawnedDisplayEntityPart.PartType.MANNEQUIN) return;
-        setAndSend(DisplayAttributes.Mannequin.SCALE, scale);
+        setAndSend(DisplayAttributes.Mannequin.SCALE, (float) scale);
     }
 
     @Override
